// dicta/src/lib/generated/UpdateStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UpdateInfo } from './UpdateInfo'
import type { UpdateProgress } from './UpdateProgress'

export type UpdateStatus =
  | { status: 'checking' }
  | { status: 'available'; update: UpdateInfo }
  | { status: 'notAvailable' }
  | { status: 'downloading'; progress: UpdateProgress }
  | { status: 'downloaded' }
  | { status: 'installing' }
  | { status: 'error'; message: string }